Re-factored Android build scripts. Bumped GameActivity to 2.0.0-alpha01.
