+++ code/demo/android/app/src/main/java/org/orx/demo/MainActivity.java
@@ -1,4 +1,4 @@
-package org.orxproject.orxtest;
+package org.orx.demo;
 
 import org.orx.lib.OrxGameActivity;
