+++ code/demo/android/app/src/main/jni/Android.mk
@@ -1,19 +1,18 @@
 LOCAL_PATH := $(call my-dir)
-SRC_PATH := $(LOCAL_PATH)
-
-ifeq ($(strip $(ORX)),)
-$(error ORX environment variable not set)
-endif
 
 include $(CLEAR_VARS)
 
 LOCAL_MODULE := orxTest
-LOCAL_SRC_FILES := $(SRC_PATH)/orxTest.cpp
-LOCAL_STATIC_LIBRARIES := orx
+LOCAL_SRC_FILES := orxTest.cpp
+LOCAL_CFLAGS := -DNO_MALLINFO=1 -DTARGET_OS_ANDROID
+LOCAL_STATIC_LIBRARIES := orx cpufeatures
+
+LOCAL_LDFLAGS += -v
 
 LOCAL_ARM_MODE := arm
 
 include $(BUILD_SHARED_LIBRARY)
 
-$(call import-add-path,$(ORX))
-$(call import-module,build/android/jni)
+$(call import-module,prefab/orx)
+
+$(call import-module,android/cpufeatures)
