+++ code/demo/android/app/src/main/jni/orxTest.cpp
@@ -31,7 +31,6 @@
  *
  */
 
-
 #include "orx.h"
 
 static orxOBJECT* spstGenerator;
