+++ code/demo/android/gradle/wrapper/gradle-wrapper.properties
@@ -1,4 +1,4 @@
-#Tue Dec 27 23:54:45 CET 2022
+#Sun Jan 15 01:46:56 CET 2023
 distributionBase=GRADLE_USER_HOME
 distributionUrl=https\://services.gradle.org/distributions/gradle-7.5-bin.zip
 distributionPath=wrapper/dists
